# Protocols/v1/generated/angular-client/model/nodeCollection.ts
/**
 * PolarisApi
 * This is the api for Polaris Data Analysis Project on  [PolarisGithub](https://github.com/Star-Academy/StarAcademy-Group2/) 
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Node } from './node';

/**
 * An array of nodes
 */
export interface NodeCollection extends Array<Node> { 
}