# Protocols/v1/generated/asp-server/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "PolarisApi",
    "description" : "This is the api for Polaris Data Analysis Project on \n[PolarisGithub](https://github.com/Star-Academy/StarAcademy-Group2/)\n",
    "license" : {
      "name" : "MIT License",
      "url" : "https://github.com/Star-Academy/StarAcademy-Group1/blob/master/LICENSE"
    },
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost:5000/api/v1",
    "description" : "HTTP local host"
  }, {
    "url" : "https://localhost:5001/api/v1",
    "description" : "HTTPS local host"
  } ],
...
  "paths" : {
...
    "/nodes/{nodeId}" : {
      "get" : {
        "tags" : [ "nodes" ],
        "summary" : "Get Nodes by id",
        "description" : "Returns a single Node",
        "operationId" : "getNodeById",
        "parameters" : [ {
          "name" : "nodeId",
          "in" : "path",
          "description" : "Id of the Node to return",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Node"
                }
              }
            }
          },
          "404" : {
            "description" : "Node not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "nodes" ],
        "summary" : "Delete Nodes by id",
        "operationId" : "deleteNodeById",
        "parameters" : [ {
          "name" : "nodeId",
          "in" : "path",
          "description" : "Node id to delete",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "success"
          },
          "404" : {
            "description" : "node not found"
          }
        }
      }
    },
    "/nodes" : {
      "get" : {
        "tags" : [ "nodes" ],
        "summary" : "Get Node(s)",
        "operationId" : "getNodes",
        "parameters" : [ {
          "name" : "filter",
          "in" : "query",
          "description" : "Filter on Nodes in [\"{field} {operator} {value}\", ...] format",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Graph"
          }
        }, {
          "name" : "pageIndex",
          "in" : "query",
          "description" : "Pagination pageIndex",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "Pagination pageSize",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/NodeCollection"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "nodes" ],
        "summary" : "Update an existing Node",
        "operationId" : "updateNode",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/Node"
        },
        "responses" : {
          "200" : {
            "description" : "Updated"
          },
          "404" : {
            "description" : "Node not found"
          },
          "405" : {
            "description" : "Invalid input"
          }
        }
      },
      "post" : {
        "tags" : [ "nodes" ],
        "summary" : "Add a new Node to the dataset",
        "operationId" : "addNode",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/Node"
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "405" : {
            "description" : "Invalid input"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Node" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "Unique id of the node"
          },
          "data" : {
            "type" : "object",
            "description" : "Data of the entity that is set on the node"
          }
        },
        "description" : "Node of the graph",
        "example" : {
          "id" : "6000000198",
          "data" : {
            "cardId" : "6104340000000000",
            "sheba" : "IR120778801496000000198",
            "accountType" : "پس انداز",
            "branchTelephone" : "55638667",
            "branchAdress" : "تهران-خیابان خیام-بالاتر از چهارراه گلوبندک",
            "branchName" : "گلوبندک",
            "ownerName" : "افسر",
            "ownerFamilyName" : "طباطبایی",
            "ownerId" : "1227114110"
          }
        }
      },
      "Edge" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "Unique id of the edge"
          },
          "source" : {
            "type" : "string",
            "description" : "Unique id of the source node"
          },
          "target" : {
            "type" : "string",
            "description" : "Unique id of the target node"
          },
          "data" : {
            "type" : "object",
            "description" : "Data of the entity that is set on the edge"
          }
        },
        "description" : "Edge of the graph",
        "example" : {
          "id" : "38791925753",
          "source" : "3000000271",
          "target" : "7000000434",
          "data" : {
            "timestamp" : "2018-05-24 6:42:05",
            "trackingId" : 661520,
            "amount" : 8370000,
            "type" : "کارت به کارت"
          }
        }
      },
      "NodeCollection" : {
        "type" : "array",
        "description" : "An array of nodes",
        "items" : {
          "$ref" : "#/components/schemas/Node"
        }
      },
      "EdgeCollection" : {
        "type" : "array",
        "description" : "An array of edges",
        "items" : {
          "$ref" : "#/components/schemas/Edge"
        }
      },
      "Graph" : {
        "type" : "object",
        "properties" : {
          "nodes" : {
            "$ref" : "#/components/schemas/NodeCollection"
          },
          "edges" : {
            "$ref" : "#/components/schemas/EdgeCollection"
          }
        },
        "description" : "A Set containing unique nodes and unique edges",
        "example" : {
          "nodes" : [ {
            "id" : "6000000198",
            "data" : {
              "cardId" : "6104340000000000",
              "sheba" : "IR120778801496000000198",
              "accountType" : "پس انداز",
              "branchTelephone" : "55638667",
              "branchAdress" : "تهران-خیابان خیام-بالاتر از چهارراه گلوبندک",
              "branchName" : "گلوبندک",
              "ownerName" : "افسر",
              "ownerFamilyName" : "طباطبایی",
              "ownerId" : "1227114110"
            }
          }, {
            "id" : "6000000198",
            "data" : {
              "cardId" : "6104340000000000",
              "sheba" : "IR120778801496000000198",
              "accountType" : "پس انداز",
              "branchTelephone" : "55638667",
              "branchAdress" : "تهران-خیابان خیام-بالاتر از چهارراه گلوبندک",
              "branchName" : "گلوبندک",
              "ownerName" : "افسر",
              "ownerFamilyName" : "طباطبایی",
              "ownerId" : "1227114110"
            }
          } ],
          "edges" : [ {
            "id" : "38791925753",
            "source" : "3000000271",
            "target" : "7000000434",
            "data" : {
              "timestamp" : "2018-05-24 6:42:05",
              "trackingId" : 661520,
              "amount" : 8370000,
              "type" : "کارت به کارت"
            }
          }, {
            "id" : "38791925753",
            "source" : "3000000271",
            "target" : "7000000434",
            "data" : {
              "timestamp" : "2018-05-24 6:42:05",
              "trackingId" : 661520,
              "amount" : 8370000,
              "type" : "کارت به کارت"
            }
          } ]
        }
      },
...
    },
    "requestBodies" : {
      "Node" : {
        "description" : "Node object to be sent",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Node"
            }
          }
        },
        "required" : true
      },
...
    }
  }
}